backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-rest-framework==0.1.0
django-split-settings==1.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
weasyprint==58.1
djoser==2.1.0
drf-extra-fields==3.4.1
django-filter==23.1